requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
pydantic-settings
sqlalchemy
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-dotenv
alembic
aiosmtplib>=11.0.0
# §4.5 - §4.6 Módulos IA
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
SpeechRecognition>=3.10.0
joblib>=1.3.0
aiofiles>=23.0.0
